+++ src/core/component/functional/helpers.ts
@@ -76,8 +76,8 @@ export function inheritContext(
 
 	// Here, the functional component is recreated during re-rendering.
 	// Therefore, the destructor call should not recursively propagate to child components.
-	// Also, we should not call the unmount in the directives used within the component.
-	parentCtx.unsafe.$destroy(<ComponentDestructorOptions>{recursive: false, callUnmount: false});
+	// Also, we should not unmount the vnodes created within the component.
+	parentCtx.unsafe.$destroy(<ComponentDestructorOptions>{recursive: false, shouldUnmountVNodes: false});
 
 	const
 		props = ctx.$props,